+++ database/migrations/2021_12_10_020650_create_design_units_table.php
@@ -15,7 +15,7 @@ class CreateDesignUnitsTable extends Migration
     {
         Schema::create('designUnits', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('contrat_id')->unsigned();
+            $table->bigInteger('contrat_id')->unsigned();
             $table->double('locationBase')->nullable();
             $table->double('conducteur')->nullable();
             $table->double('siegeBebe')->nullable();
